refactor compTable
